+++ business_requirement_deliverable/models/__init__.py
@@ -1,2 +1,4 @@
 from . import business
+from . import business_requirement_deliverable
+from . import business_requirement_resource
 from . import res_partner
